requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
opencv-python==4.5.5.64
pillow==9.0.0
numpy==1.21.5
pandas==1.4.2
scikit-learn==1.0.2
flask-sqlalchemy==2.5.1
flask-login==0.5.0 